+++ Sources/SyntaxKit/PublicTypes.swift
@@ -52,9 +52,9 @@ public struct Rule: Equatable, Sendable {
         self.match = match
         self.begin = begin
         self.end = end
-        self.captures = captures.sorted { $0.index < $1.index }
-        self.beginCaptures = beginCaptures.sorted { $0.index < $1.index }
-        self.endCaptures = endCaptures.sorted { $0.index < $1.index }
+        self.captures = captures
+        self.beginCaptures = beginCaptures
+        self.endCaptures = endCaptures
         self.include = include
         self.patterns = patterns
     }